CMake: Remove misleading opts add DLL support
This commit fixes #114

The options UNICODE and BUILD_DYNAMIC_MODULE were only changing the name
of the libraries and not the flags used for compilation.
I removed them for now, but if we want to add Unicode/Runtime version to
CMake it will be possible later.
This fixes #114 as the .lib created by CMake on windows can no more be
mistaken for another one. If people need to build a specific version of
the library, they can for now use the cmake cache to change the /M(D/T)
flag and/or add the UNICODE and _UNICODE defines.

I changed the option name for the test executable so that it won't
conflict with other projects, and made sure BUILD_TESTING is checked.
When BUILD_TESTING (the official option for CTest) is ON,
STORM_BUILD_TESTS is by default ON. If BUILD_TESTING is set to OFF, then
STORM_BUILD_TESTS can only be set to OFF.

I also added the stormlib_dll/StormLib.def file to the source list when
using MSVC, which means you can now correctly create StormLib.dll and
its associated .lib and .exp by setting BUILD_SHARED_LIBS to ON.