Set SFileSetFilePointer error when new pointer < 0
This addresses an apparent oversight in the following commit:

- 951f416398b3aa0b32969b4a391a3103443ad99e

An error is not actually set, even though the return value now correctly
indicates failure.  Per the documentation for [`SetFilerPointer ()`][1],
this should be set to `ERROR_NEGATIVE_SEEK` on Windows.  On Mac/Linux,
this would be `EINVAL` as per the documentation for [`fseek ()`][2].

[1]: https://msdn.microsoft.com/en-us/library/windows/desktop/aa365541(v=vs.85).aspx
[2]: http://pubs.opengroup.org/onlinepubs/9699919799/functions/fseek.html